Move script modules into utils folder
